+++ .gitignore
@@ -1,5 +1,6 @@
-/cpp/*.o
-/cpp/*.efi
+*/*.o
+*/*.efi
 /mnt
 /target
 /edk2
+.vscode/
