+++ cpp/main.cpp
@@ -9,21 +9,21 @@ typedef EFI_STATUS (*EFI_TEXT_STRING)(
 
 typedef struct _EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL
 {
-  void *dummy;
-  EFI_TEXT_STRING OutputString;
+    void *dummy;
+    EFI_TEXT_STRING OutputString;
 } EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL;
 
 typedef struct
 {
-  char dummy[52];
-  EFI_HANDLE ConsoleOutHandle;
-  EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL *ConOut;
+    char dummy[52];
+    EFI_HANDLE ConsoleOutHandle;
+    EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL *ConOut;
 } EFI_SYSTEM_TABLE;
 
 EFI_STATUS EfiMain(EFI_HANDLE ImageHandle, EFI_SYSTEM_TABLE *SystemTable)
 {
-  SystemTable->ConOut->OutputString(SystemTable->ConOut, (CHAR16 *)L"Hello, World\n");
-  while (1)
-    ;
-  return 0;
+    SystemTable->ConOut->OutputString(SystemTable->ConOut, (CHAR16 *)L"Hello, World\n");
+    while (1)
+        ;
+    return 0;
 }
