+++ src/main.rs
@@ -2,361 +2,61 @@
 #![no_main]
 #![feature(offset_of)]
 
-use core::cmp::min;
-use core::fmt;
-use core::mem::offset_of;
-use core::mem::size_of;
+use core::fmt::Write;
 use core::panic::PanicInfo;
-use core::ptr::null_mut;
+use core::writeln;
 
-type EfiVoid = u8;
-type EfiHandle = u64;
-type Result<T> = core::result::Result<T, &'static str>;
-
-#[repr(C)]
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-struct EfiGuid {
-    pub data0: u32,
-    pub data1: u16,
-    pub data2: u16,
-    pub data3: [u8; 8],
-}
-
-const EFI_GRAPHICS_OUTPUT_PROTOCOL_GUID: EfiGuid = EfiGuid {
-    data0: 0x9042a9de,
-    data1: 0x23dc,
-    data2: 0x4a38,
-    data3: [0x96, 0xfb, 0x7a, 0xde, 0xd0, 0x80, 0x51, 0x6a],
-};
-
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-#[must_use]
-#[repr(u64)]
-enum EfiStatus {
-    Success = 0,
-}
-
-#[repr(i64)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[allow(non_camel_case_types)]
-pub enum EfiMemoryType {
-    RESERVED = 0,
-    LOADER_CODE,
-    LOADER_DATA,
-    BOOT_SERVICES_CODE,
-    BOOT_SERVICES_DATA,
-    RUNTIME_SERVICES_CODE,
-    RUNTIME_SERVICES_DATA,
-    CONVENTIONAL_MEMORY,
-    UNUSABLE_MEMORY,
-    ACPI_RECLAIM_MEMORY,
-    ACPI_MEMORY_NVS,
-    MEMORY_MAPPED_IO,
-    MEMORY_MAPPED_IO_PORT_SPACE,
-    PAL_CODE,
-    PERSITENT_MEMORY,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-struct EfiMemoryDescriptor {
-    memory_type: EfiMemoryType,
-    phisical_start: u64,
-    virtual_start: u64,
-    number_of_pages: u64,
-    attribute: u64,
-}
-const MEMORY_MAP_BUFFER_SIZE: usize = 0x8000;
-
-struct MemoryMapHolder {
-    memory_map_buffer: [u8; MEMORY_MAP_BUFFER_SIZE],
-    memory_map_size: usize,
-    map_key: usize,
-    descriptor_size: usize,
-    descriptor_version: u32,
-}
-
-struct MemoryMapIterator<'a> {
-    map: &'a MemoryMapHolder,
-    ofs: usize,
-}
-
-impl<'a> Iterator for MemoryMapIterator<'a> {
-    type Item = &'a EfiMemoryDescriptor;
-    fn next(&mut self) -> Option<&'a EfiMemoryDescriptor> {
-        if self.ofs >= self.map.memory_map_size {
-            None
-        } else {
-            let e: &EfiMemoryDescriptor = unsafe {
-                &*(self.map.memory_map_buffer.as_ptr().add(self.ofs) as *const &EfiMemoryDescriptor)
-            };
-            self.ofs += self.map.descriptor_size;
-            Some(e)
-        }
-    }
-}
-
-impl MemoryMapHolder {
-    pub const fn new() -> MemoryMapHolder {
-        MemoryMapHolder {
-            memory_map_buffer: [0; MEMORY_MAP_BUFFER_SIZE],
-            memory_map_size: MEMORY_MAP_BUFFER_SIZE,
-            map_key: 0,
-            descriptor_size: 0,
-            descriptor_version: 0,
-        }
-    }
-}
-
-#[repr(C)]
-struct EfiBootServiciesTable {
-    _reserved0: [u64; 7],
-    get_memory_map: extern "win64" fn(
-        memory_map_size: *mut usize,
-        memory_map: *mut u8,
-        map_key: *mut usize,
-        descriptor_size: *mut usize,
-        descriptor_version: *mut u32,
-    ) -> EfiStatus,
-    _reserved1: [u64; 32],
-    locate_protocol: extern "win64" fn(
-        protocol: *const EfiGuid,
-        registration: *const EfiVoid,
-        interface: *mut *mut EfiVoid,
-    ) -> EfiStatus,
-}
-
-impl EfiBootServiciesTable {
-    fn get_memory_map(&self, map: &mut MemoryMapHolder) -> EfiStatus {
-        (self.get_memory_map)(
-            &mut map.memory_map_size,
-            map.memory_map_buffer.as_mut_ptr(),
-            &mut map.map_key,
-            &mut map.descriptor_size,
-            &mut map.descriptor_version,
-        )
-    }
-}
-const _: () = assert!(offset_of!(EfiBootServiciesTable, get_memory_map) == 56);
-const _: () = assert!(offset_of!(EfiBootServiciesTable, locate_protocol) == 320);
-
-#[repr(C)]
-struct EfiSystemTable {
-    _reserved0: [u64; 12],
-    pub boot_servicies: &'static EfiBootServiciesTable,
-}
-const _: () = assert!(offset_of!(EfiSystemTable, boot_servicies) == 96);
-
-#[repr(C)]
-#[derive(Debug)]
-struct EfiGraphicsOutputProtocolPixelInfo {
-    version: u32,
-    pub horizontal_resolution: u32,
-    pub vertial_resolution: u32,
-    _padding0: [u32; 5],
-    pub pixels_per_scan_line: u32,
-}
-const _: () = assert!(size_of::<EfiGraphicsOutputProtocolPixelInfo>() == 36);
-
-#[repr(C)]
-#[derive(Debug)]
-struct EfiGraphicsOutputProtocolMode<'a> {
-    pub mac_mode: u32,
-    pub mode: u32,
-    pub info: &'a EfiGraphicsOutputProtocolPixelInfo,
-    pub size_of_info: u64,
-    pub frame_buffer_base: usize,
-    pub frame_buffer_size: usize,
-}
-
-#[repr(C)]
-#[derive(Debug)]
-struct EfiGraphicsOutputProtocol<'a> {
-    reserved: [u64; 3],
-    pub mode: &'a EfiGraphicsOutputProtocolMode<'a>,
-}
-
-fn locate_graphic_protocol<'a>(
-    efi_system_table: &EfiSystemTable,
-) -> Result<&'a EfiGraphicsOutputProtocol<'a>> {
-    let mut graphic_output_protocol = null_mut::<EfiGraphicsOutputProtocol>();
-    let status = (efi_system_table.boot_servicies.locate_protocol)(
-        &EFI_GRAPHICS_OUTPUT_PROTOCOL_GUID,
-        null_mut::<EfiVoid>(),
-        &mut graphic_output_protocol as *mut *mut EfiGraphicsOutputProtocol as *mut *mut EfiVoid,
-    );
-    if status != EfiStatus::Success {
-        return Err("Failed to locate graphics output protocol");
-    }
-    Ok(unsafe { &*graphic_output_protocol })
-}
-
-use core::arch::asm;
-fn hlt() {
-    unsafe { asm!("hlt") }
-}
+use micro_os::graphics::fill_rect;
+use micro_os::graphics::Bitmap;
+use micro_os::qemu::exit_qemu;
+use micro_os::qemu::QemuExitCode;
+use micro_os::uefi::exit_from_efi_services;
+use micro_os::uefi::init_vram;
+use micro_os::uefi::EfiHandle;
+use micro_os::uefi::EfiMemoryType;
+use micro_os::uefi::EfiSystemTable;
+use micro_os::uefi::MemoryMapHolder;
+use micro_os::uefi::VramTextWriter;
+use micro_os::x86::hlt;
 
 #[no_mangle]
-fn efi_main(_image_handle: EfiHandle, efi_system_table: &EfiSystemTable) {
+fn efi_main(image_handle: EfiHandle, efi_system_table: &EfiSystemTable) {
     let mut vram = init_vram(efi_system_table).expect("init_vra, failed");
-    for y in 0..vram.height {
-        for x in 0..vram.width {
-            if let Some(pixel) = vram.pixel_at_mut(x, y) {
-                *pixel = 0x000000;
-            }
+
+    let vw = vram.width();
+    let vh = vram.height();
+    fill_rect(&mut vram, 0x0000ff, 0, 0, vw, vh).expect("fill_rect failed");
+
+    let mut w = VramTextWriter::new(&mut vram);
+    let mut memory_map = MemoryMapHolder::new();
+
+    let status = efi_system_table
+        .boot_servicies()
+        .get_memory_map(&mut memory_map);
+    writeln!(w, "{status:?}").unwrap();
+    let mut total_memory_pages = 0;
+    for e in memory_map.iter() {
+        if e.memory_type() != EfiMemoryType::CONVENTIONAL_MEMORY {
+            continue;
         }
+        total_memory_pages += e.number_of_pages();
+        writeln!(w, "{e:?}").unwrap();
     }
+    let total_memory_size_mib = total_memory_pages * 4096 / 1024 / 1024;
+    writeln!(
+        w,
+        "Total: {total_memory_pages} pages = {total_memory_size_mib}MiB"
+    )
+    .unwrap();
 
-    // let mut w = VramTextWriter::new(&mut vram);
-
-    // for i in 0..4 {
-    //     writeln!(w, "i = {i}").unwrap();
-    // }
-    // let mut memory_map = MemoryMapHolder::new();
-    // let status = efi_system_table
-    //     .boot_servicies
-    //     .get_memory_map(&mut memory_map);
-    draw_str_fg(&mut vram, 0, 0, 0xffffff, "Hello World!");
+    exit_from_efi_services(image_handle, efi_system_table, &mut memory_map);
+    writeln!(w, "Hello, Non-UEFI World!").unwrap();
     loop {
         hlt();
     }
-}
-
-trait Bitmap {
-    fn bytes_per_pixel(&self) -> i64;
-    fn pixles_per_line(&self) -> i64;
-    fn width(&self) -> i64;
-    fn height(&self) -> i64;
-    fn buf_mut(&mut self) -> *mut u8;
-
-    unsafe fn unchecked_pixled_at_mut(&mut self, x: i64, y: i64) -> *mut u32 {
-        self.buf_mut()
-            .add(((y * self.pixles_per_line() + x) * self.bytes_per_pixel()) as usize)
-            as *mut u32
-    }
-
-    fn pixel_at_mut(&mut self, x: i64, y: i64) -> Option<&mut u32> {
-        if self.is_in_x_range(x) && self.is_in_y_range(y) {
-            unsafe { Some(&mut *(self.unchecked_pixled_at_mut(x, y))) }
-        } else {
-            None
-        }
-    }
-
-    fn is_in_x_range(&self, px: i64) -> bool {
-        0 <= px && px < min(self.width(), self.pixles_per_line())
-    }
-    fn is_in_y_range(&self, py: i64) -> bool {
-        0 <= py && py < self.height()
-    }
-}
-
-#[derive(Clone, Copy)]
-struct VramBufferInfo {
-    buf: *mut u8,
-    width: i64,
-    height: i64,
-    pixels_per_line: i64,
-}
-
-impl Bitmap for VramBufferInfo {
-    fn bytes_per_pixel(&self) -> i64 {
-        4
-    }
-    fn pixles_per_line(&self) -> i64 {
-        self.pixels_per_line
-    }
-    fn width(&self) -> i64 {
-        self.width
-    }
-    fn height(&self) -> i64 {
-        self.height
-    }
-    fn buf_mut(&mut self) -> *mut u8 {
-        self.buf
-    }
-}
-
-fn init_vram(efi_system_table: &EfiSystemTable) -> Result<VramBufferInfo> {
-    let gp = locate_graphic_protocol(efi_system_table)?;
-    Ok(VramBufferInfo {
-        buf: gp.mode.frame_buffer_base as *mut u8,
-        width: gp.mode.info.horizontal_resolution as i64,
-        height: gp.mode.info.vertial_resolution as i64,
-        pixels_per_line: gp.mode.info.pixels_per_scan_line as i64,
-    })
-}
-
-fn draw_str_fg<T: Bitmap>(buf: &mut T, x: i64, y: i64, color: u32, s: &str) {
-    for (i, c) in s.chars().enumerate() {
-        draw_font_fg(buf, i as i64 * 16 + 256, 0, 0xffffff, c);
-    }
-}
-
-fn draw_font_fg<T: Bitmap>(buf: &mut T, x: i64, y: i64, color: u32, c: char) {
-    if let Some(font) = lookup_font(c) {
-        for (dy, raw) in font.iter().enumerate() {
-            for (dx, pixel) in raw.iter().enumerate() {
-                let color = match pixel {
-                    '*' => color,
-                    _ => continue,
-                };
-                let _ = draw_point(buf, color, x + dx as i64, y + dy as i64);
-            }
-        }
-    }
-}
-
-fn draw_point<T: Bitmap>(buf: &mut T, color: u32, x: i64, y: i64) -> Result<()> {
-    *(buf.pixel_at_mut(x, y).ok_or("Out of Range")?) = color;
-    Ok(())
-}
-
-struct VramTextWriter<'a> {
-    vram: &'a mut VramBufferInfo,
-}
-
-impl<'a> VramTextWriter<'a> {
-    fn new(vram: &'a mut VramBufferInfo) -> Self {
-        Self { vram }
-    }
-}
-
-impl fmt::Write for VramTextWriter<'_> {
-    fn write_str(&mut self, s: &str) -> fmt::Result {
-        draw_str_fg(self.vram, 0, 0, 0xffffff, s);
-        Ok(())
-    }
-}
-
-fn lookup_font(c: char) -> Option<[[char; 8]; 16]> {
-    const FONT_SOURCE: &str = include_str!("../font/font.txt");
-    if let Ok(c) = u8::try_from(c) {
-        let mut fi = FONT_SOURCE.split('\n');
-        while let Some(line) = fi.next() {
-            if let Some(line) = line.strip_prefix("0x") {
-                if let Ok(idx) = u8::from_str_radix(line, 16) {
-                    if idx != c {
-                        continue;
-                    }
-                    let mut font = [['*'; 8]; 16];
-                    for (y, line) in fi.clone().take(16).enumerate() {
-                        for (x, c) in line.chars().enumerate() {
-                            if let Some(e) = font[y].get_mut(x) {
-                                *e = c;
-                            }
-                        }
-                    }
-                    return Some(font);
-                }
-            }
-        }
-    }
-    None
 }
 
 #[panic_handler]
-fn panic(_panic: &PanicInfo<'_>) -> ! {
-    loop {
-        hlt();
-    }
+fn panic(_info: &PanicInfo) -> ! {
+    exit_qemu(QemuExitCode::Fail);
 }
