+++ src/result.rs
@@ -1,0 +1,1 @@
+pub type Result<T> = core::result::Result<T, &'static str>;
